Add: ✨ Nested Object features for input field as an example
This is useful to group some data on it's own seperated object.
The reason could be that the API accepts the data in such a manner,
or the database stirs the data in a particular fashion.

Whatever might be reason, we want to group together some of the fields
in our form by using nested objects