Add: ✨ Field Level validation
This validate prop is available on the fast Field component as well.

Why would you use field validation over form level validation?

This depends on your project. One possible use case would be that
you want to render your fields based on a json that you fetch from an
API call. Once you get the JSON representation of your form, you
iterate over the different objects building the validation function in
each iteration and assigning it to the validate prop on the field or
the fast Field component