Add: ✨ Manually triggering validation to check if the field is valid or not through form "top-level"
Why do we have the same props at the form level and the field level?

Use the field level object when you have to deal with that individual field, but if there
is something  that has to be done for the entire form, then use the form level formik props

A simple manual way to think of field level validation is to check if a username already
exist in the database. You can check a username button which will perform that validation