Add: ✨ Add form validation check automatically for all field values when it is populated on the site by using "dirty" object
"dirty" object from formik a boolean value which indicates if at least one of
the form fields value has changed since it was initialized

The drawback of this solution where the object of the initialValues could contain values inside it.
That means the submit button is still "disabled". Even you as a developer knows that the state value
is valid. Just because you haven't changed the form field value, and not allowed to submit the form
which is something is not good.

What you can think of this solution of using "dirty" is to disable the
submit button is based on the assumption that on page load that is without
the user changing any of the form field values. The form state is always
invalid. If you know for a fact the user will interact with your form
and enter values which will never be exactly the same values as the
initial values object, then you can stick to this option.

Although this seems like a really bad option. The user will always interact
with the form without trying to click on submit, so this is definently a good option
to use as "dirty"